fix: moved persist functions
