+++ packages/bruno-app/src/components/CodeEditor/index.js
@@ -6,7 +6,7 @@
  */
 
 import React, { createRef } from 'react';
-import { isEqual, escapeRegExp } from 'lodash';
+import { isEqual } from 'lodash';
 import { defineCodeMirrorBrunoVariablesMode } from 'utils/common/codemirror';
 import { setupAutoComplete, showRootHints } from 'utils/codemirror/autocomplete';
 import StyledWrapper from './StyledWrapper';
@@ -17,6 +17,14 @@ import { getAllVariables } from 'utils/collections';
 import { setupLinkAware } from 'utils/codemirror/linkAware';
 import { setupLintErrorTooltip } from 'utils/codemirror/lint-errors';
 import CodeMirrorSearch from 'components/CodeMirrorSearch/index';
+import {
+  applyEditorState,
+  captureEditorState,
+  getDocKey,
+  isTabRecentlyCleared,
+  readPersistedEditorState,
+  writePersistedEditorState
+} from './state-persistence';
 
 const CodeMirror = require('codemirror');
 window.jsonlint = jsonlint;
@@ -24,110 +32,7 @@ window.JSHINT = JSHINT;
 
 const TAB_SIZE = 2;
 
-/*
- * Per-tab editor state persistence — why this exists:
- *
- * Every tab switch causes CodeMirror's setValue() to wipe folds, cursor,
- * selection, undo history, and scroll position. To preserve them, we serialize
- * the relevant pieces to localStorage under a stable key for each editor and
- * re-apply them on mount / tab switch. CodeMirror exposes a JSON-serializable
- * representation of its undo stack via getHistory()/setHistory(), which is what
- * makes Cmd-Z continue working across switches.
- *
- * Note: we deliberately do NOT persist the content itself — the canonical value
- * lives in Redux (props.value). We only persist the editor's "view" state on
- * top of that content. If content has drifted between save and restore, fold
- * positions are applied leniently (foldCode silently no-ops on invalid lines)
- * and history is skipped to avoid an inconsistent undo stack.
- */
-const STORAGE_PREFIX = 'persisted::codeeditor::';
-
-const readPersistedEditorState = (key) => {
-  try {
-    const raw = localStorage.getItem(STORAGE_PREFIX + key);
-    return raw ? JSON.parse(raw) : null;
-  } catch {
-    return null;
-  }
-};
-
-const writePersistedEditorState = (key, state) => {
-  try {
-    if (state == null) {
-      localStorage.removeItem(STORAGE_PREFIX + key);
-    } else {
-      localStorage.setItem(STORAGE_PREFIX + key, JSON.stringify(state));
-    }
-  } catch {
-    // localStorage may be unavailable or full — silently ignore
-  }
-};
-
-// Removes every persisted CodeEditor entry whose docKey starts with `${tabUid}:`.
-// Called from the closeTabs thunk so a closed tab doesn't leak its editor state
-// (folds, history, cursor, scroll) in localStorage forever.
-export const clearCodeEditorPersistedState = (tabUid) => {
-  if (!tabUid) return;
-  try {
-    const prefix = `${STORAGE_PREFIX}${tabUid}:`;
-    Object.keys(localStorage)
-      .filter((k) => k.startsWith(prefix))
-      .forEach((k) => localStorage.removeItem(k));
-  } catch {
-    // localStorage may be unavailable — silently ignore
-  }
-};
-
-const captureEditorState = (editor) => {
-  if (!editor) return null;
-  const doc = editor.getDoc();
-  const folds = editor
-    .getAllMarks()
-    .filter((m) => m.__isFold)
-    .map((m) => m.find())
-    .filter(Boolean)
-    .map((range) => range.from);
-  return {
-    contentLength: doc.getValue().length,
-    cursor: doc.getCursor(),
-    selections: doc.listSelections(),
-    history: doc.getHistory(),
-    folds,
-    scrollY: editor.getScrollInfo().top
-  };
-};
-
-const applyEditorState = (editor, state, currentContent) => {
-  if (!editor || !state) return;
-  const doc = editor.getDoc();
-  const contentMatches = state.contentLength === (currentContent || '').length;
-
-  // History/cursor/selection only make sense if content didn't drift — applying
-  // a stale undo stack to different content would let Cmd-Z replay edits that
-  // no longer correspond to anything visible.
-  if (contentMatches) {
-    if (state.history) {
-      try { doc.setHistory(state.history); } catch {}
-    }
-    if (state.cursor) {
-      try { doc.setCursor(state.cursor); } catch {}
-    }
-    if (state.selections && state.selections.length) {
-      try { doc.setSelections(state.selections); } catch {}
-    }
-  }
-  // Folds are cheap and lenient — try them either way.
-  if (state.folds && state.folds.length) {
-    editor.operation(() => {
-      state.folds.forEach((from) => {
-        try { editor.foldCode(from); } catch {}
-      });
-    });
-  }
-  if (state.scrollY != null) {
-    try { editor.scrollTo(null, state.scrollY); } catch {}
-  }
-};
+export { clearCodeEditorPersistedState } from './state-persistence';
 
 export default class CodeEditor extends React.Component {
   constructor(props) {
@@ -153,22 +58,10 @@ export default class CodeEditor extends React.Component {
     };
   }
 
-  // Identifies which Doc in docCache belongs to this CodeEditor instance.
-  //
-  // Callers can pass an explicit `docKey` prop when the auto-derived key would
-  // collide — e.g. Pre-Request vs Post-Response script editors share the same
-  // item/mode/readOnly and need an extra disambiguator.
-  //
-  // Auto-derived parts:
-  //   id       — distinguishes different tabs (requests or collections)
-  //   mode     — distinguishes editors within the same tab (e.g. JSON body vs JS script)
-  //   readOnly — distinguishes response viewer (ro) from body editor (rw) when modes match
+  // Thin wrapper around the pure getDocKey helper from state-persistence.js.
+  // Kept on the class so the rest of the lifecycle code reads naturally.
   _getDocKey() {
-    if (this.props.docKey) return this.props.docKey;
-    const id = this.props.item?.uid || this.props.collection?.uid || 'default';
-    const mode = this.props.mode || 'default';
-    const readOnly = this.props.readOnly ? 'ro' : 'rw';
-    return `${id}:${mode}:${readOnly}`;
+    return getDocKey(this.props);
   }
 
   componentDidMount() {
@@ -439,7 +332,12 @@ export default class CodeEditor extends React.Component {
       // Snapshot view state to localStorage before tearing down the editor so
       // the next mount of a CodeEditor with this docKey can restore folds,
       // cursor, selection, undo history, and scroll position.
-      if (this._currentDocKey) {
+      //
+      // Skip the write if this tab is currently being closed — the closeTabs
+      // thunk has just cleared this tab's entries and we'd otherwise re-write
+      // them right back, defeating the cleanup.
+      const tabUid = this.props.item?.uid || this.props.collection?.uid;
+      if (this._currentDocKey && !isTabRecentlyCleared(tabUid)) {
         writePersistedEditorState(this._currentDocKey, captureEditorState(this.editor));
       }
